\name{streamFPM-package}
\alias{streamFPM-package}
\alias{streamFPM}
\docType{package}
\title{
Stream Frequent Pattern Mining
}
\description{
Algorithms and data generators for mining for frequent patterns on data streams
}
\details{
\tabular{ll}{
Package: \tab streamFPM\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-01-05\cr
%License: \tab What license is it under?\cr
}
streamFPM expansions of the functionality of stream. It introduces several Frequent Pattern Mining Algorithms that can be used on streams of Transaction data.
}
\author{
Derek Phanekham <ds.phanekham@smu.edu>

}
\references{
%add literature here FIXME
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%~~ simple examples of the most important functions ~~
}
